+++ buildSrc/src/main/kotlin/deps.kt
@@ -96,4 +96,5 @@ object Dependencies {
   const val checkStyle = "com.puppycrawl.tools:checkstyle:${Versions.checkStyle}"
   const val signatureAndroid21 = "net.sf.androidscents.signature:android-api-level-21:5.0.1_r2@signature"
   const val signatureJava18 = "org.codehaus.mojo.signature:java18:1.0@signature"
+  const val assertk = "com.willowtreeapps.assertk:assertk:0.25"
 }
