+++ okhttp/src/jvmMain/kotlin/okhttp3/Headers.kt
@@ -24,35 +24,34 @@ import java.util.Date
 import java.util.Locale
 import java.util.TreeMap
 import java.util.TreeSet
-import okhttp3.Headers.Builder
-import okhttp3.internal.format
+import okhttp3.internal.commonAdd
+import okhttp3.internal.commonAddAll
+import okhttp3.internal.commonAddLenient
+import okhttp3.internal.commonBuild
+import okhttp3.internal.commonEquals
+import okhttp3.internal.commonGet
+import okhttp3.internal.commonHashCode
+import okhttp3.internal.commonHeadersGet
+import okhttp3.internal.commonHeadersOf
+import okhttp3.internal.commonIterator
+import okhttp3.internal.commonName
+import okhttp3.internal.commonNewBuilder
+import okhttp3.internal.commonRemoveAll
+import okhttp3.internal.commonSet
+import okhttp3.internal.commonToHeaders
+import okhttp3.internal.commonToString
+import okhttp3.internal.commonValue
+import okhttp3.internal.commonValues
+import okhttp3.internal.headersCheckName
 import okhttp3.internal.http.toHttpDateOrNull
 import okhttp3.internal.http.toHttpDateString
-import okhttp3.internal.isSensitiveHeader
 import org.codehaus.mojo.animal_sniffer.IgnoreJRERequirement
 
-/**
- * The header fields of a single HTTP message. Values are uninterpreted strings; use `Request` and
- * `Response` for interpreted headers. This class maintains the order of the header fields within
- * the HTTP message.
- *
- * This class tracks header values line-by-line. A field with multiple comma- separated values on
- * the same line will be treated as a field with a single value by this class. It is the caller's
- * responsibility to detect and split on commas if their field permits multiple values. This
- * simplifies use of single-valued fields whose values routinely contain commas, such as cookies or
- * dates.
- *
- * This class trims whitespace from values. It never returns values with leading or trailing
- * whitespace.
- *
- * Instances of this class are immutable. Use [Builder] to create instances.
- */
 @Suppress("NAME_SHADOWING")
-class Headers private constructor(
-  private val namesAndValues: Array<String>
+actual class Headers internal actual constructor(
+  internal actual val namesAndValues: Array<String>
 ) : Iterable<Pair<String, String>> {
-  /** Returns the last value corresponding to the specified field, or null. */
-  operator fun get(name: String): String? = get(namesAndValues, name)
+  actual operator fun get(name: String): String? = commonHeadersGet(namesAndValues, name)
 
   /**
    * Returns the last value corresponding to the specified field parsed as an HTTP date, or null if
@@ -70,8 +69,7 @@ class Headers private constructor(
     return value?.toInstant()
   }
 
-  /** Returns the number of field values. */
-  @get:JvmName("size") val size: Int
+  @get:JvmName("size") actual val size: Int
     get() = namesAndValues.size / 2
 
   @JvmName("-deprecated_size")
@@ -81,14 +79,11 @@ class Headers private constructor(
       level = DeprecationLevel.ERROR)
   fun size(): Int = size
 
-  /** Returns the field at `position`. */
-  fun name(index: Int): String = namesAndValues[index * 2]
+  actual fun name(index: Int): String = commonName(index)
 
-  /** Returns the value at `index`. */
-  fun value(index: Int): String = namesAndValues[index * 2 + 1]
+  actual fun value(index: Int): String = commonValue(index)
 
-  /** Returns an immutable case-insensitive set of header names. */
-  fun names(): Set<String> {
+  actual fun names(): Set<String> {
     val result = TreeSet(String.CASE_INSENSITIVE_ORDER)
     for (i in 0 until size) {
       result.add(name(i))
@@ -96,21 +91,7 @@ class Headers private constructor(
     return Collections.unmodifiableSet(result)
   }
 
-  /** Returns an immutable list of the header values for `name`. */
-  fun values(name: String): List<String> {
-    var result: MutableList<String>? = null
-    for (i in 0 until size) {
-      if (name.equals(name(i), ignoreCase = true)) {
-        if (result == null) result = ArrayList(2)
-        result.add(value(i))
-      }
-    }
-    return if (result != null) {
-      Collections.unmodifiableList(result)
-    } else {
-      emptyList()
-    }
-  }
+  actual fun values(name: String): List<String> = commonValues(name)
 
   /**
    * Returns the number of bytes required to encode these headers using HTTP/1.1. This is also the
@@ -129,83 +110,15 @@ class Headers private constructor(
     return result
   }
 
-  override operator fun iterator(): Iterator<Pair<String, String>> {
-    return Array(size) { name(it) to value(it) }.iterator()
-  }
+  actual override operator fun iterator(): Iterator<Pair<String, String>> = commonIterator()
 
-  fun newBuilder(): Builder {
-    val result = Builder()
-    result.namesAndValues += namesAndValues
-    return result
-  }
+  actual fun newBuilder(): Builder = commonNewBuilder()
 
-  /**
-   * Returns true if `other` is a `Headers` object with the same headers, with the same casing, in
-   * the same order. Note that two headers instances may be *semantically* equal but not equal
-   * according to this method. In particular, none of the following sets of headers are equal
-   * according to this method:
-   *
-   * 1. Original
-   * ```
-   * Content-Type: text/html
-   * Content-Length: 50
-   * ```
-   *
-   * 2. Different order
-   *
-   * ```
-   * Content-Length: 50
-   * Content-Type: text/html
-   * ```
-   *
-   * 3. Different case
-   *
-   * ```
-   * content-type: text/html
-   * content-length: 50
-   * ```
-   *
-   * 4. Different values
-   *
-   * ```
-   * Content-Type: text/html
-   * Content-Length: 050
-   * ```
-   *
-   * Applications that require semantically equal headers should convert them into a canonical form
-   * before comparing them for equality.
-   */
-  override fun equals(other: Any?): Boolean {
-    return other is Headers && namesAndValues.contentEquals(other.namesAndValues)
-  }
+  actual override fun equals(other: Any?): Boolean = commonEquals(other)
 
-  override fun hashCode(): Int = namesAndValues.contentHashCode()
+  override fun hashCode(): Int = commonHashCode()
 
-  /**
-   * Returns header names and values. The names and values are separated by `: ` and each pair is
-   * followed by a newline character `\n`.
-   *
-   * Since OkHttp 5 this redacts these sensitive headers:
-   *
-   *  * `Authorization`
-   *  * `Cookie`
-   *  * `Proxy-Authorization`
-   *  * `Set-Cookie`
-   *
-   * To get all headers as a human-readable string use `toMultimap().toString()`.
-   */
-  override fun toString(): String {
-    return buildString {
-      for (i in 0 until size) {
-        val name = name(i)
-        val value = value(i)
-        append(name)
-        append(": ")
-        append(if (isSensitiveHeader(name)) "██" else value)
-        append("\n")
-      }
-    }
-  }
+  actual override fun toString(): String = commonToString()
 
   fun toMultimap(): Map<String, List<String>> {
     val result = TreeMap<String, MutableList<String>>(String.CASE_INSENSITIVE_ORDER)
@@ -221,8 +134,8 @@ class Headers private constructor(
     return result
   }
 
-  class Builder {
-    internal val namesAndValues: MutableList<String> = ArrayList(20)
+  actual class Builder {
+    internal actual val namesAndValues: MutableList<String> = ArrayList(20)
 
     /**
      * Add a header line without any validation. Only appropriate for headers from the remote peer
@@ -253,32 +166,18 @@ class Headers private constructor(
       add(line.substring(0, index).trim(), line.substring(index + 1))
     }
 
-    /**
-     * Add a header with the specified name and value. Does validation of header names and values.
-     */
-    fun add(name: String, value: String) = apply {
-      checkName(name)
-      checkValue(value, name)
-      addLenient(name, value)
-    }
+    actual fun add(name: String, value: String) = commonAdd(name, value)
 
     /**
      * Add a header with the specified name and value. Does validation of header names, allowing
      * non-ASCII values.
      */
     fun addUnsafeNonAscii(name: String, value: String) = apply {
-      checkName(name)
+      headersCheckName(name)
       addLenient(name, value)
     }
 
-    /**
-     * Adds all headers from an existing collection.
-     */
-    fun addAll(headers: Headers) = apply {
-      for (i in 0 until headers.size) {
-        addLenient(headers.name(i), headers.value(i))
-      }
-    }
+    actual fun addAll(headers: Headers) = commonAddAll(headers)
 
     /**
      * Add a header with the specified name and formatted date. Does validation of header names and
@@ -318,83 +217,26 @@ class Headers private constructor(
      * Add a field with the specified value without any validation. Only appropriate for headers
      * from the remote peer or cache.
      */
-    internal fun addLenient(name: String, value: String) = apply {
-      namesAndValues.add(name)
-      namesAndValues.add(value.trim())
-    }
+    internal fun addLenient(name: String, value: String) = commonAddLenient(name, value)
 
-    fun removeAll(name: String) = apply {
-      var i = 0
-      while (i < namesAndValues.size) {
-        if (name.equals(namesAndValues[i], ignoreCase = true)) {
-          namesAndValues.removeAt(i) // name
-          namesAndValues.removeAt(i) // value
-          i -= 2
-        }
-        i += 2
-      }
-    }
+    actual fun removeAll(name: String) = commonRemoveAll(name)
 
     /**
      * Set a field with the specified value. If the field is not found, it is added. If the field is
      * found, the existing values are replaced.
      */
-    operator fun set(name: String, value: String) = apply {
-      checkName(name)
-      checkValue(value, name)
-      removeAll(name)
-      addLenient(name, value)
-    }
+    actual operator fun set(name: String, value: String) = commonSet(name, value)
 
     /** Equivalent to `build().get(name)`, but potentially faster. */
-    operator fun get(name: String): String? {
-      for (i in namesAndValues.size - 2 downTo 0 step 2) {
-        if (name.equals(namesAndValues[i], ignoreCase = true)) {
-          return namesAndValues[i + 1]
-        }
-      }
-      return null
-    }
+    actual operator fun get(name: String): String? = commonGet(name)
 
-    fun build(): Headers = Headers(namesAndValues.toTypedArray())
+    actual fun build(): Headers = commonBuild()
   }
 
-  companion object {
-    private fun get(namesAndValues: Array<String>, name: String): String? {
-      for (i in namesAndValues.size - 2 downTo 0 step 2) {
-        if (name.equals(namesAndValues[i], ignoreCase = true)) {
-          return namesAndValues[i + 1]
-        }
-      }
-      return null
-    }
-
-    /**
-     * Returns headers for the alternating header names and values. There must be an even number of
-     * arguments, and they must alternate between header names and values.
-     */
+  actual companion object {
     @JvmStatic
     @JvmName("of")
-    fun headersOf(vararg namesAndValues: String): Headers {
-      require(namesAndValues.size % 2 == 0) { "Expected alternating header names and values" }
-
-      // Make a defensive copy and clean it up.
-      val namesAndValues: Array<String> = namesAndValues.clone() as Array<String>
-      for (i in namesAndValues.indices) {
-        require(namesAndValues[i] != null) { "Headers cannot be null" }
-        namesAndValues[i] = namesAndValues[i].trim()
-      }
-
-      // Check for malformed headers.
-      for (i in namesAndValues.indices step 2) {
-        val name = namesAndValues[i]
-        val value = namesAndValues[i + 1]
-        checkName(name)
-        checkValue(value, name)
-      }
-
-      return Headers(namesAndValues)
-    }
+    actual fun headersOf(vararg namesAndValues: String): Headers = commonHeadersOf(*namesAndValues)
 
     @JvmName("-deprecated_of")
     @Deprecated(
@@ -405,25 +247,9 @@ class Headers private constructor(
       return headersOf(*namesAndValues)
     }
 
-    /** Returns headers for the header names and values in the [Map]. */
     @JvmStatic
     @JvmName("of")
-    fun Map<String, String>.toHeaders(): Headers {
-      // Make a defensive copy and clean it up.
-      val namesAndValues = arrayOfNulls<String>(size * 2)
-      var i = 0
-      for ((k, v) in this) {
-        val name = k.trim()
-        val value = v.trim()
-        checkName(name)
-        checkValue(value, name)
-        namesAndValues[i] = name
-        namesAndValues[i + 1] = value
-        i += 2
-      }
-
-      return Headers(namesAndValues as Array<String>)
-    }
+    actual fun Map<String, String>.toHeaders(): Headers = commonToHeaders()
 
     @JvmName("-deprecated_of")
     @Deprecated(
@@ -432,26 +258,6 @@ class Headers private constructor(
         level = DeprecationLevel.ERROR)
     fun of(headers: Map<String, String>): Headers {
       return headers.toHeaders()
-    }
-
-    private fun checkName(name: String) {
-      require(name.isNotEmpty()) { "name is empty" }
-      for (i in name.indices) {
-        val c = name[i]
-        require(c in '\u0021'..'\u007e') {
-          format("Unexpected char %#04x at %d in header name: %s", c.code, i, name)
-        }
-      }
-    }
-
-    private fun checkValue(value: String, name: String) {
-      for (i in value.indices) {
-        val c = value[i]
-        require(c == '\t' || c in '\u0020'..'\u007e') {
-          format("Unexpected char %#04x at %d in %s value", c.code, i, name) +
-              (if (isSensitiveHeader(name)) "" else ": $value")
-        }
-      }
     }
   }
 }
