+++ okhttp/src/jvmTest/java/okhttp3/HeadersChallengesTest.kt
@@ -15,466 +15,19 @@
  */
 package okhttp3
 
-import java.time.Instant
-import java.util.Date
-import okhttp3.Headers.Companion.headersOf
-import okhttp3.Headers.Companion.toHeaders
-import okhttp3.TestUtil.headerEntries
-import okhttp3.internal.EMPTY_HEADERS
 import okhttp3.internal.http.parseChallenges
-import okhttp3.internal.http2.Http2ExchangeCodec.Companion.http2HeadersList
-import okhttp3.internal.http2.Http2ExchangeCodec.Companion.readHttp2HeadersList
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Assertions.fail
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 
-class HeadersTest {
-  @Test fun readNameValueBlockDropsForbiddenHeadersHttp2() {
-    val headerBlock = headersOf(
-      ":status", "200 OK",
-      ":version", "HTTP/1.1",
-      "connection", "close"
-    )
-    val request = Request.Builder().url("http://square.com/").build()
-    val response = readHttp2HeadersList(headerBlock, Protocol.HTTP_2).request(request).build()
-    val headers = response.headers
-    assertThat(headers.size).isEqualTo(1)
-    assertThat(headers.name(0)).isEqualTo(":version")
-    assertThat(headers.value(0)).isEqualTo("HTTP/1.1")
-  }
-
-  @Test fun http2HeadersListDropsForbiddenHeadersHttp2() {
-    val request = Request.Builder()
-      .url("http://square.com/")
-      .header("Connection", "upgrade")
-      .header("Upgrade", "websocket")
-      .header("Host", "square.com")
-      .header("TE", "gzip")
-      .build()
-    val expected = headerEntries(
-      ":method", "GET",
-      ":path", "/",
-      ":authority", "square.com",
-      ":scheme", "http"
-    )
-    assertThat(http2HeadersList(request)).isEqualTo(expected)
-  }
-
-  @Test fun http2HeadersListDontDropTeIfTrailersHttp2() {
-    val request = Request.Builder()
-      .url("http://square.com/")
-      .header("TE", "trailers")
-      .build()
-    val expected = headerEntries(
-      ":method", "GET",
-      ":path", "/",
-      ":scheme", "http",
-      "te", "trailers"
-    )
-    assertThat(http2HeadersList(request)).isEqualTo(expected)
-  }
-
-  @Test fun ofTrims() {
-    val headers = headersOf("\t User-Agent \n", " \r OkHttp ")
-    assertThat(headers.name(0)).isEqualTo("User-Agent")
-    assertThat(headers.value(0)).isEqualTo("OkHttp")
-  }
-
-  @Test fun addParsing() {
-    val headers = Headers.Builder()
-      .add("foo: bar")
-      .add(" foo: baz") // Name leading whitespace is trimmed.
-      .add("foo : bak") // Name trailing whitespace is trimmed.
-      .add("\tkey\t:\tvalue\t") // '\t' also counts as whitespace
-      .add("ping:  pong  ") // Value whitespace is trimmed.
-      .add("kit:kat") // Space after colon is not required.
-      .build()
-    assertThat(headers.values("foo")).containsExactly("bar", "baz", "bak")
-    assertThat(headers.values("key")).containsExactly("value")
-    assertThat(headers.values("ping")).containsExactly("pong")
-    assertThat(headers.values("kit")).containsExactly("kat")
-  }
-
-  @Test fun addThrowsOnEmptyName() {
-    try {
-      Headers.Builder().add(": bar")
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-    try {
-      Headers.Builder().add(" : bar")
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun addThrowsOnNoColon() {
-    try {
-      Headers.Builder().add("foo bar")
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun addThrowsOnMultiColon() {
-    try {
-      Headers.Builder().add(":status: 200 OK")
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun addUnsafeNonAsciiRejectsUnicodeName() {
-    try {
-      Headers.Builder()
-        .addUnsafeNonAscii("héader1", "value1")
-        .build()
-      fail<Any>("Should have complained about invalid value")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message).isEqualTo("Unexpected char 0xe9 at 1 in header name: héader1")
-    }
-  }
-
-  @Test fun addUnsafeNonAsciiAcceptsUnicodeValue() {
-    val headers = Headers.Builder()
-      .addUnsafeNonAscii("header1", "valué1")
-      .build()
-    assertThat(headers.toString()).isEqualTo("header1: valué1\n")
-  }
-
-  @Test fun ofThrowsOddNumberOfHeaders() {
-    try {
-      headersOf("User-Agent", "OkHttp", "Content-Length")
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun ofThrowsOnEmptyName() {
-    try {
-      headersOf("", "OkHttp")
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun ofAcceptsEmptyValue() {
-    val headers = headersOf("User-Agent", "")
-    assertThat(headers.value(0)).isEqualTo("")
-  }
-
-  @Test fun ofMakesDefensiveCopy() {
-    val namesAndValues = arrayOf(
-      "User-Agent",
-      "OkHttp"
-    )
-    val headers = headersOf(*namesAndValues)
-    namesAndValues[1] = "Chrome"
-    assertThat(headers.value(0)).isEqualTo("OkHttp")
-  }
-
-  @Test fun ofRejectsNullChar() {
-    try {
-      headersOf("User-Agent", "Square\u0000OkHttp")
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun ofMapThrowsOnNull() {
-    try {
-      (mapOf("User-Agent" to null) as Map<String, String>).toHeaders()
-      fail<Any>()
-    } catch (expected: NullPointerException) {
-    }
-  }
-
-  @Test fun ofMapThrowsOnEmptyName() {
-    try {
-      mapOf("" to "OkHttp").toHeaders()
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun ofMapThrowsOnBlankName() {
-    try {
-      mapOf(" " to "OkHttp").toHeaders()
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun ofMapAcceptsEmptyValue() {
-    val headers = mapOf("User-Agent" to "").toHeaders()
-    assertThat(headers.value(0)).isEqualTo("")
-  }
-
-  @Test fun ofMapTrimsKey() {
-    val headers = mapOf(" User-Agent " to "OkHttp").toHeaders()
-    assertThat(headers.name(0)).isEqualTo("User-Agent")
-  }
-
-  @Test fun ofMapTrimsValue() {
-    val headers = mapOf("User-Agent" to " OkHttp ").toHeaders()
-    assertThat(headers.value(0)).isEqualTo("OkHttp")
-  }
-
-  @Test fun ofMapMakesDefensiveCopy() {
-    val namesAndValues = mutableMapOf<String, String>()
-    namesAndValues["User-Agent"] = "OkHttp"
-    val headers = namesAndValues.toHeaders()
-    namesAndValues["User-Agent"] = "Chrome"
-    assertThat(headers.value(0)).isEqualTo("OkHttp")
-  }
-
-  @Test fun ofMapRejectsNullCharInName() {
-    try {
-      mapOf("User-\u0000Agent" to "OkHttp").toHeaders()
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun ofMapRejectsNullCharInValue() {
-    try {
-      mapOf("User-Agent" to "Square\u0000OkHttp").toHeaders()
-      fail<Any>()
-    } catch (expected: IllegalArgumentException) {
-    }
-  }
-
-  @Test fun toMultimapGroupsHeaders() {
-    val headers = headersOf(
-      "cache-control", "no-cache",
-      "cache-control", "no-store",
-      "user-agent", "OkHttp"
-    )
-    val headerMap = headers.toMultimap()
-    assertThat(headerMap["cache-control"]!!.size).isEqualTo(2)
-    assertThat(headerMap["user-agent"]!!.size).isEqualTo(1)
-  }
-
-  @Test fun toMultimapUsesCanonicalCase() {
-    val headers = headersOf(
-      "cache-control", "no-store",
-      "Cache-Control", "no-cache",
-      "User-Agent", "OkHttp"
-    )
-    val headerMap = headers.toMultimap()
-    assertThat(headerMap["cache-control"]!!.size).isEqualTo(2)
-    assertThat(headerMap["user-agent"]!!.size).isEqualTo(1)
-  }
-
-  @Test fun toMultimapAllowsCaseInsensitiveGet() {
-    val headers = headersOf(
-      "cache-control", "no-store",
-      "Cache-Control", "no-cache"
-    )
-    val headerMap = headers.toMultimap()
-    assertThat(headerMap["cache-control"]!!.size).isEqualTo(2)
-    assertThat(headerMap["Cache-Control"]!!.size).isEqualTo(2)
-  }
-
-  @Test fun nameIndexesAreStrict() {
-    val headers = headersOf("a", "b", "c", "d")
-    try {
-      headers.name(-1)
-      fail<Any>()
-    } catch (expected: IndexOutOfBoundsException) {
-    }
-    assertThat(headers.name(0)).isEqualTo("a")
-    assertThat(headers.name(1)).isEqualTo("c")
-    try {
-      headers.name(2)
-      fail<Any>()
-    } catch (expected: IndexOutOfBoundsException) {
-    }
-  }
-
-  @Test fun valueIndexesAreStrict() {
-    val headers = headersOf("a", "b", "c", "d")
-    try {
-      headers.value(-1)
-      fail<Any>()
-    } catch (expected: IndexOutOfBoundsException) {
-    }
-    assertThat(headers.value(0)).isEqualTo("b")
-    assertThat(headers.value(1)).isEqualTo("d")
-    try {
-      headers.value(2)
-      fail<Any>()
-    } catch (expected: IndexOutOfBoundsException) {
-    }
-  }
-
-  @Test fun builderRejectsUnicodeInHeaderName() {
-    try {
-      Headers.Builder().add("héader1", "value1")
-      fail<Any>("Should have complained about invalid name")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 1 in header name: héader1")
-    }
-  }
-
-  @Test fun builderRejectsUnicodeInHeaderValue() {
-    try {
-      Headers.Builder().add("header1", "valué1")
-      fail<Any>("Should have complained about invalid value")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 4 in header1 value: valué1")
-    }
-  }
-
-  @Test fun varargFactoryRejectsUnicodeInHeaderName() {
-    try {
-      headersOf("héader1", "value1")
-      fail<Any>("Should have complained about invalid value")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 1 in header name: héader1")
-    }
-  }
-
-  @Test fun varargFactoryRejectsUnicodeInHeaderValue() {
-    try {
-      headersOf("header1", "valué1")
-      fail<Any>("Should have complained about invalid value")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 4 in header1 value: valué1")
-    }
-  }
-
-  @Test fun mapFactoryRejectsUnicodeInHeaderName() {
-    try {
-      mapOf("héader1" to "value1").toHeaders()
-      fail<Any>("Should have complained about invalid value")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 1 in header name: héader1")
-    }
-  }
-
-  @Test fun mapFactoryRejectsUnicodeInHeaderValue() {
-    try {
-      mapOf("header1" to "valué1").toHeaders()
-      fail<Any>("Should have complained about invalid value")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 4 in header1 value: valué1")
-    }
-  }
-
-  @Test fun sensitiveHeadersNotIncludedInExceptions() {
-    try {
-      Headers.Builder().add("Authorization", "valué1")
-      fail<Any>("Should have complained about invalid name")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 4 in Authorization value")
-    }
-    try {
-      Headers.Builder().add("Cookie", "valué1")
-      fail<Any>("Should have complained about invalid name")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 4 in Cookie value")
-    }
-    try {
-      Headers.Builder().add("Proxy-Authorization", "valué1")
-      fail<Any>("Should have complained about invalid name")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 4 in Proxy-Authorization value")
-    }
-    try {
-      Headers.Builder().add("Set-Cookie", "valué1")
-      fail<Any>("Should have complained about invalid name")
-    } catch (expected: IllegalArgumentException) {
-      assertThat(expected.message)
-        .isEqualTo("Unexpected char 0xe9 at 4 in Set-Cookie value")
-    }
-  }
-
-  @Test fun headersEquals() {
-    val headers1 = Headers.Builder()
-      .add("Connection", "close")
-      .add("Transfer-Encoding", "chunked")
-      .build()
-    val headers2 = Headers.Builder()
-      .add("Connection", "close")
-      .add("Transfer-Encoding", "chunked")
-      .build()
-    assertThat(headers2).isEqualTo(headers1)
-    assertThat(headers2.hashCode()).isEqualTo(headers1.hashCode())
-  }
-
-  @Test fun headersNotEquals() {
-    val headers1 = Headers.Builder()
-      .add("Connection", "close")
-      .add("Transfer-Encoding", "chunked")
-      .build()
-    val headers2 = Headers.Builder()
-      .add("Connection", "keep-alive")
-      .add("Transfer-Encoding", "chunked")
-      .build()
-    assertThat(headers2).isNotEqualTo(headers1)
-    assertThat(headers2.hashCode()).isNotEqualTo(headers1.hashCode().toLong())
-  }
-
-  @Test fun headersToString() {
-    val headers = Headers.Builder()
-      .add("A", "a")
-      .add("B", "bb")
-      .build()
-    assertThat(headers.toString()).isEqualTo("A: a\nB: bb\n")
-  }
-
-  @Test fun headersToStringRedactsSensitiveHeaders() {
-    val headers = Headers.Builder()
-      .add("content-length", "99")
-      .add("authorization", "peanutbutter")
-      .add("proxy-authorization", "chocolate")
-      .add("cookie", "drink=coffee")
-      .add("set-cookie", "accessory=sugar")
-      .add("user-agent", "OkHttp")
-      .build()
-    assertThat(headers.toString()).isEqualTo(
-      """
-      |content-length: 99
-      |authorization: ██
-      |proxy-authorization: ██
-      |cookie: ██
-      |set-cookie: ██
-      |user-agent: OkHttp
-      |""".trimMargin()
-    )
-  }
-
-  @Test fun headersAddAll() {
-    val sourceHeaders = Headers.Builder()
-      .add("A", "aa")
-      .add("a", "aa")
-      .add("B", "bb")
-      .build()
-    val headers = Headers.Builder()
-      .add("A", "a")
-      .addAll(sourceHeaders)
-      .add("C", "c")
-      .build()
-    assertThat(headers.toString()).isEqualTo("A: a\nA: aa\na: aa\nB: bb\nC: c\n")
-  }
+class HeadersChallengesTest {
 
   /** See https://github.com/square/okhttp/issues/2780.  */
   @Test fun testDigestChallengeWithStrictRfc2617Header() {
     val headers = Headers.Builder()
       .add(
         "WWW-Authenticate", "Digest realm=\"myrealm\", nonce=\"fjalskdflwejrlaskdfjlaskdjflaks"
-        + "jdflkasdf\", qop=\"auth\", stale=\"FALSE\""
+          + "jdflkasdf\", qop=\"auth\", stale=\"FALSE\""
       )
       .build()
     val challenges = headers.parseChallenges("WWW-Authenticate")
@@ -493,7 +46,7 @@ class HeadersTest {
     val headers = Headers.Builder()
       .add(
         "WWW-Authenticate", "Digest qop=\"auth\", realm=\"myrealm\", nonce=\"fjalskdflwejrlask"
-        + "dfjlaskdjflaksjdflkasdf\", stale=\"FALSE\""
+          + "dfjlaskdjflaksjdflkasdf\", stale=\"FALSE\""
       )
       .build()
     val challenges = headers.parseChallenges("WWW-Authenticate")
@@ -512,7 +65,7 @@ class HeadersTest {
     val headers = Headers.Builder()
       .add(
         "WWW-Authenticate", "Digest qop=\"auth\", nonce=\"fjalskdflwejrlaskdfjlaskdjflaksjdflk"
-        + "asdf\", realm=\"myrealm\", stale=\"FALSE\""
+          + "asdf\", realm=\"myrealm\", stale=\"FALSE\""
       )
       .build()
     val challenges = headers.parseChallenges("WWW-Authenticate")
@@ -531,7 +84,7 @@ class HeadersTest {
     val headers = Headers.Builder()
       .add(
         "WWW-Authenticate", "Digest qop=\"auth\", underrealm=\"myrealm\", nonce=\"fjalskdflwej"
-        + "rlaskdfjlaskdjflaksjdflkasdf\", stale=\"FALSE\""
+          + "rlaskdfjlaskdjflaksjdflkasdf\", stale=\"FALSE\""
       )
       .build()
     val challenges = headers.parseChallenges("WWW-Authenticate")
@@ -550,7 +103,7 @@ class HeadersTest {
     val headers = Headers.Builder()
       .add(
         "WWW-Authenticate", "Digest qop=\"auth\",    realm=\"myrealm\", nonce=\"fjalskdflwejrl"
-        + "askdfjlaskdjflaksjdflkasdf\", stale=\"FALSE\""
+          + "askdfjlaskdjflaksjdflkasdf\", stale=\"FALSE\""
       )
       .build()
     val challenges = headers.parseChallenges("WWW-Authenticate")
@@ -569,7 +122,7 @@ class HeadersTest {
     val headers = Headers.Builder()
       .add(
         "WWW-Authenticate", "Digest    realm=\"myrealm\", nonce=\"fjalskdflwejrlaskdfjlaskdjfl"
-        + "aksjdflkasdf\", qop=\"auth\", stale=\"FALSE\""
+          + "aksjdflkasdf\", qop=\"auth\", stale=\"FALSE\""
       )
       .build()
     val challenges = headers.parseChallenges("WWW-Authenticate")
@@ -588,7 +141,7 @@ class HeadersTest {
     val headers = Headers.Builder()
       .add(
         "WWW-Authenticate", "DiGeSt qop=\"auth\", rEaLm=\"myrealm\", nonce=\"fjalskdflwejrlask"
-        + "dfjlaskdjflaksjdflkasdf\", stale=\"FALSE\""
+          + "dfjlaskdjflaksjdflkasdf\", stale=\"FALSE\""
       )
       .build()
     val challenges = headers.parseChallenges("WWW-Authenticate")
@@ -608,7 +161,7 @@ class HeadersTest {
     val headers = Headers.Builder()
       .add(
         "WWW-Authenticate", "DIgEsT rEaLm=\"myrealm\", nonce=\"fjalskdflwejrlaskdfjlaskdjflaks"
-        + "jdflkasdf\", qop=\"auth\", stale=\"FALSE\""
+          + "jdflkasdf\", qop=\"auth\", stale=\"FALSE\""
       )
       .build()
     val challenges = headers.parseChallenges("WWW-Authenticate")
@@ -808,7 +361,7 @@ class HeadersTest {
       .build()
     assertThat(headers.parseChallenges("WWW-Authenticate"))
       .isEqualTo(listOf(Challenge("Other", mapOf(null to "abc==")))
-    )
+      )
   }
 
   @Test fun token68AndAuthParams() {
@@ -858,60 +411,5 @@ class HeadersTest {
       Challenge("Basic", mapOf("realm" to "myrealm")),
       Challenge("Basic", mapOf("realm" to "myotherrealm"))
     )
-  }
-
-  @Test fun byteCount() {
-    assertThat(EMPTY_HEADERS.byteCount()).isEqualTo(0L)
-    assertThat(
-      Headers.Builder()
-        .add("abc", "def")
-        .build()
-        .byteCount()
-    ).isEqualTo(10L)
-    assertThat(
-      Headers.Builder()
-        .add("abc", "def")
-        .add("ghi", "jkl")
-        .build()
-        .byteCount()
-    ).isEqualTo(20L)
-  }
-
-  @Test fun addDate() {
-    val expected = Date(0L)
-    val headers = Headers.Builder()
-      .add("testDate", expected)
-      .build()
-    assertThat(headers["testDate"]).isEqualTo("Thu, 01 Jan 1970 00:00:00 GMT")
-    assertThat(headers.getDate("testDate")).isEqualTo(Date(0L))
-  }
-
-  @Test fun addInstant() {
-    val expected = Instant.ofEpochMilli(0L)
-    val headers = Headers.Builder()
-      .add("Test-Instant", expected)
-      .build()
-    assertThat(headers["Test-Instant"]).isEqualTo("Thu, 01 Jan 1970 00:00:00 GMT")
-    assertThat(headers.getInstant("Test-Instant")).isEqualTo(expected)
-  }
-
-  @Test fun setDate() {
-    val expected = Date(1000)
-    val headers = Headers.Builder()
-      .add("testDate", Date(0L))
-      .set("testDate", expected)
-      .build()
-    assertThat(headers["testDate"]).isEqualTo("Thu, 01 Jan 1970 00:00:01 GMT")
-    assertThat(headers.getDate("testDate")).isEqualTo(expected)
-  }
-
-  @Test fun setInstant() {
-    val expected = Instant.ofEpochMilli(1000L)
-    val headers = Headers.Builder()
-      .add("Test-Instant", Instant.ofEpochMilli(0L))
-      .set("Test-Instant", expected)
-      .build()
-    assertThat(headers["Test-Instant"]).isEqualTo("Thu, 01 Jan 1970 00:00:01 GMT")
-    assertThat(headers.getInstant("Test-Instant")).isEqualTo(expected)
   }
 }
